# Scroll-correlated camera configuration properties