--- a/src/features/ar.ts
+++ b/src/features/ar.ts
@@ -16,8 +16,8 @@
 import {property} from 'lit-element';
 
 import {IS_ANDROID, IS_AR_QUICKLOOK_CANDIDATE, IS_IOS, IS_IOS_CHROME, IS_IOS_SAFARI, IS_WEBXR_AR_CANDIDATE} from '../constants.js';
-import {enumerationDeserializer} from '../conversions.js';
 import ModelViewerElementBase, {$container, $renderer, $scene} from '../model-viewer-base.js';
+import {enumerationDeserializer} from '../styles/deserializers.js';
 import {Constructor, deserializeUrl} from '../utilities.js';
 
 /**