![calc2](https://user-images.githubusercontent.com/240083/64276704-68189500-cefd-11e9-8bb6-e7753ffe81dd.gif)

This change proposes the introduction of a declarative, CSS-like configuration syntax for `<model-viewer>` attributes and properties. The initial goal of this change is to enable users to correlate the camera's orbital position with the top-level scroll position on the page. The long-term goal of this change is to establish a common, intuitive, web-native syntax for describing complex attribute/property configuration schemas in `<model-viewer>` as it evolves.

**The main user-facing change is that CSS-like [`calc()`](https://developer.mozilla.org/en-US/docs/Web/CSS/calc) and [`env()`](https://developer.mozilla.org/en-US/docs/Web/CSS/env) functions can now be used when configuring `camera-orbit`.** Additionally, a single "environment variable" called `window-scroll-y` is available to be used by the `env()` function, which yields a value between `0` and `1` derived from the current global scroll position.

Previously, we used a very basic string parsing mechanism to parse `camera-orbit` values in order to support CSS-like numbers with units (e.g., `10deg`) and keywords like `auto`. This change expands the parsing facilities and layers on systems for evaluating the parsed result and observing signals that tell when the parsed result should optimally be evaluated.

A trivial usage of this new syntax looks like this:

```html
<model-viewer camera-orbit="calc(-1.5rad + env(window-scroll-y) * 4rad) 90deg auto"
              src="assets/Astronaut.glb"
              background-color="#622dcf"
              alt="A 3D model of an astronaut">
</model-viewer>
```

Some details of the new syntax include:

 - **All `calc()` algebra is supported**, including nested `calc()` terms
    - We currently do not enforce any maximum nesting depth
 - **Only supports angles and lengths**; no new unit types introduced at this time
 - **Adheres closely to CSS**, only deviating where needed
    - Everything is standard CSS with the notable exception of `window-scroll-y`
 - **No breaking changes required** to inherit the new capabilities; the new syntax is additive

## New `@style` decorator

A new `@style` decorator has been introduced in order to make it relatively easy to integrate sophisticated style parsing to any property already using the `@property` decorator from the `lit-element` module.

Use of the `@style` decorator looks like this:

```typescript
// Intrinsics describe the metrics required in order to parse foo,
// including basis values for percentages, fallback values and
// substitutes for keywords. Intrinsics can optionally be specified as
// a factory function that takes an element instance and returns an
// intrinsics object: 
const fooIntrinsics = {
  basis: [numberNode(1, 'm'), numberNode(Math.PI, 'rad')],
  keywords: {auto: [null, numberNode(200, '%')]}
};

class StyleableElement extends UpdatingElement {
  @style({intrinsics: fooIntrinsics, updateHandler: 'updateFoo'})
  @property({type: String})
  foo: string = '200cm 1rad';

  updateFoo(style: [number, number]) {
    // do something with parsed value of foo
  }
}
```

## Performance analysis

Some steps have been taken to mitigate the performance cost of parsing attribute values as an expressive CSS-like syntax. That said, the performance cost of this approach is not zero and is worth analyzing.

For the purposes of this analysis, I will be using a configuration of `<model-viewer>` that stresses the style subsystem proposed here. A test page has been crafted the produces N model viewer instances in this fashion (psuedocoded):

```javascript
for (i = 0; i < N; ++i) {
  const modelViewer = document.createElement('model-viewer');
  modelViewer.src = './examples/assets/Astronaut.glb';
  modelViewer.cameraOrbit = `calc(-${i}rad + env(window-scroll-y) * 8rad) calc(180deg - env(window-scroll-y) * 180deg + 5deg * ${i}) calc(3m - env(window-scroll-y))`;
  document.body.appendChild(modelViewer);
}
```

Note that the proposed parsing/evaluation strategy caches parsed ASTs across all elements, as well as the evaluated results of static sub-trees of the AST on a per-element basis. In this test case, every element will have a unique value for `camera-orbit`, so the parser cache will be subverted.

The test case has been run with various values of N on a Google Pixel 3 phone and a 2013 MacBook Pro in Chrome. The majority of the meaningful work occurs when updating the `cameraOrbit` property, so the `updateCameraOrbit` and `syncCameraOrbit` methods were instrumented to measure the work.

### Startup costs

The upfront parsing costs of a single element on the page is quite small in the grand scheme things, but still very significant. On a MacBook Pro, initial cost of `updateCameraOrbit` grows from `0.47ms` to `1.86ms`, and on a Pixel 3 it grows from `0.8ms` to `3.6ms`.

Branch | MacBook Pro (N=1) | Pixel 3 (N=1)
--------|--------------------|-------------
`master` | ![image](https://user-images.githubusercontent.com/240083/64358207-e5eea600-cfba-11e9-8302-74cf63056352.png)|![image](https://user-images.githubusercontent.com/240083/64358255-f99a0c80-cfba-11e9-9587-2008b07f2ea0.png)
`scroll-correlated-camera-orbit`|![image](https://user-images.githubusercontent.com/240083/64357632-e3d81780-cfb9-11e9-9eb8-239bfc339fe2.png)|![image](https://user-images.githubusercontent.com/240083/64357589-cefb8400-cfb9-11e9-836e-755728be5495.png)

While this time is clearly dwarfed by the time we spend preparing and rendering a 3D scene, we should consider these times in the context of when they will matter most. In the future, we will probably lazy load most of our rendering implementation so that 3D content can defer to other, more critical content on the page. In this future, we will have compartmentalized most of the cost of this element, but will still ask users to pay the cost of attribute/property parsing up front. If we apply this parsing scheme to other properties as well, it seems likely that the parsing / evaluation time will start to add up.

I also checked the scaling properties of this boot-up cost when multiple elements are on the page. The total cost scales linearly with the number of elements on the page, which makes sense given the stated conditions of this test.

MacBook Pro | Pixel 3
--------------|-------
![image](https://user-images.githubusercontent.com/240083/64359129-98733880-cfbc-11e9-8878-0dc2e129dd90.png)|![image](https://user-images.githubusercontent.com/240083/64359148-a2953700-cfbc-11e9-90b5-8c0b97d3bc2c.png)

Notably, the cost of `updateCamerOrbit` for N=1 is significantly higher than subsequent instances, which suggests that some significant amount of cost is paid upfront that I haven't quite put my finger on yet. It could be the cost of adding a listener for the scroll event (which only happens once for all elements), but that seems unlikely to explain all of the extra time. In this related data for the MacBook Pro, note the average time for N=1 compared to other sample counts (similar results were seen on the Pixel 3):

Element count | Sample size | Average time (ms) | Total time (ms)
-- | -- | -- | --
1 | 3 | 1.12167 | 3.365
5 | 15 | 0.38867 | 5.83
10 | 30 | 0.298 | 8.94
30 | 90 | 0.27767 | 24.99
100 | 300 | 0.37107 | 111.32

### Scroll synchronization costs

Additionally, I did some testing of the cost to synchronize scroll in real time. The cost to synchronize to scroll proved to be very small, so I only ran some light tests on my MacBook Pro. In these tests, I created a bunch of elements and then scrolled down the page to the bottom while collecting samples:

Method | Element count | Sample size | Average time (ms) | Total time (ms)
-- | -- | -- | -- | --
updateCameraOrbit | 10 | 30 | 0.32183 | 9.655
syncCameraOrbit | 10 | 2370 | 0.04226 | 100.15
updateCameraOrbit | 100 | 300 | 0.40555 | 121.665
syncCameraOrbit | 100 | 49600 | 0.31537 | 15642.15

We can see that the cost of `syncCameraOrbit`, which represents the cost of evaluating the AST (including incorporating dynamic scroll position values on each frame) is quite small compared to `updateCameraOrbit` (which must also parse configuration values). But, the `updateCameraOrbit` cost is a one-time startup cost that only needs to be paid when `camera-orbit` is assigned directly.

## Additional thoughts

There are a few reasons why we probably want to introduce a syntax like this to our attribute configurations:

**It gives us the flexibility to keep our API surface concise.** As time goes on, our configuration API must strike a balance between allowing complex attribute values and proliferating attribute names.

**It leverages existing user know-how.** CSS is a lingua franca of the web. Since this syntax carries with it most of the semantics of CSS, our users don't require special knowledge in order to use it.

**It scales well to different use cases.** CSS expressions can be as simple as a word (e.g., `blue`), but they also allow you to express lists, typed numbers, functions and bindings.

**It supports our goal to explore and experiment in this domain.** Over time, a syntax like this will inform our conception of a CSS-based style API as it might apply to a `<model>` tag that is built-in to the browser.

I hope that over time we will begin to express other attributes such as `field-of-view`, `camera-target` and perhaps many others in terms of the syntax afforded by this change.

Fixes #600 
Fixes #828